Fixed an error in the rule generation.
Rules of the same colour which were separated by empty blocks would not have the correct split or group status.

Any cell which is not brushed when iterating over the cells will cause the prevColor to be set to null to force a new temp rule on the next brushed cell.